+++ src/test/kotlin/blackjack/domain/player/DealerTest.kt
@@ -1,4 +1,4 @@
-package blackjack.domain.user
+package blackjack.domain.player
 
 import blackjack.domain.card.Ace
 import blackjack.domain.card.CardType
